Enable choosing Earth/Mars start dates
r4394
2018-07-06

Note 1 : this commit enables changing the Mars/Earth start dates based
         on discussions at issue
https://github.com/mars-sim/mars-sim/issues/23
	     about Mars timekeeping. We select the year 2028 arbitrarily
		 as the beginning of the year 0 of Mars Calendar. This may
         or may not have to do with the day of the first human landing
         on Mars.

Note 2 : Incidentially, according to
http://planetary.org/explore/space-topics/mars/mars-calendar.html
         the date of Aug 17 2028 coincides with L_s = 0, the beginning
		 of the Spring equinox for the northern hemisphere on Mars.

# CHANGES
1. Add earthStartDate and marsStartDate in SimulationConfig.
2. Compute the Temporarily designate the start date pair for Mars/Earth
   as follows in simulation.xml :
   a. <earth-start-date-time value="2028-08-17 15:23:13.740"/>
   b. <mars-start-date-time value="0000-Adir-01:000.000" />

Note 3 : we hand pick the time on Earth at 15:23:13.740 while the time
         on Mars is at 000.000 millisols and Ls = 0.1991 (nearly zero).
         Ordinarilly, mission control would choose the first landing at
         the time of the day with sufficient daylight. Will post more
		 timekeeping info on Mars Calendar wiki at
https://github.com/mars-sim/mars-sim/wiki/Mars-Calendar

3. Implement ClockListener in MainScene and AudioPlayer for pause
   control.
4. Switch all float param to double in sound related classes for
   consistency purpose.

# FIX
1. Fix refreshing the labels in Time Tool for both classic and JavaFX
   mode.
   - Use both ui pulse and clock pulse to update various labels and
     calendar.

+++ mars-sim-ui/src/main/java/org/mars_sim/msp/ui/swing/MainWindowMenu.java
@@ -268,8 +268,8 @@ public class MainWindowMenu extends JMenuBar implements ActionListener, MenuList
 		// Create Volume slider menu item
 		// MainDesktopPane desktop = mainWindow.getDesktop();
 		final AudioPlayer soundPlayer = desktop.getSoundPlayer();
-		float volume = soundPlayer.getMusicVolume();
-		int intVolume = Math.round(volume * 10F);
+		double volume = soundPlayer.getMusicVolume();
+		int intVolume = (int) Math.round(volume * 10.0);
 
 		musicVolumeItem = new JSliderMW(JSlider.HORIZONTAL, 0, 10, intVolume); // $NON-NLS-1$
 		musicVolumeItem.setMajorTickSpacing(1);
@@ -466,22 +466,22 @@ public class MainWindowMenu extends JMenuBar implements ActionListener, MenuList
 		}
 
 		else if (selectedItem == musicVolumeUpItem) {
-			float oldvolume = desktop.getSoundPlayer().getMusicVolume();
+			double oldvolume = desktop.getSoundPlayer().getMusicVolume();
 			desktop.getSoundPlayer().setMusicVolume(oldvolume + 0.05F);
 		}
 
 		else if (selectedItem == musicVolumeDownItem) {
-			float oldvolume = desktop.getSoundPlayer().getMusicVolume();
+			double oldvolume = desktop.getSoundPlayer().getMusicVolume();
 			desktop.getSoundPlayer().setMusicVolume(oldvolume - 0.05F);
 		}
 
 		else if (selectedItem == effectVolumeUpItem) {
-			float oldvolume = desktop.getSoundPlayer().getEffectVolume();
+			double oldvolume = desktop.getSoundPlayer().getEffectVolume();
 			desktop.getSoundPlayer().setSoundVolume(oldvolume + 0.05F);
 		}
 
 		else if (selectedItem == effectVolumeDownItem) {
-			float oldvolume = desktop.getSoundPlayer().getEffectVolume();
+			double oldvolume = desktop.getSoundPlayer().getEffectVolume();
 			desktop.getSoundPlayer().setSoundVolume(oldvolume - 0.05F);
 		}
 
@@ -537,9 +537,9 @@ public class MainWindowMenu extends JMenuBar implements ActionListener, MenuList
 		showUnitBarItem.setSelected(desktop.getMainWindow().getUnitToolBar().isVisible());
 		showToolBarItem.setSelected(desktop.getMainWindow().getToolToolBar().isVisible());
 
-		musicVolumeItem.setValue(Math.round(desktop.getSoundPlayer().getMusicVolume() * 10F));
+		musicVolumeItem.setValue((int) Math.round(desktop.getSoundPlayer().getMusicVolume() * 10));
 		musicVolumeItem.setEnabled(!desktop.getSoundPlayer().isMusicMute());
-		effectVolumeItem.setValue(Math.round(desktop.getSoundPlayer().getEffectVolume() * 10F));
+		effectVolumeItem.setValue((int) Math.round(desktop.getSoundPlayer().getEffectVolume() * 10));
 		effectVolumeItem.setEnabled(!desktop.getSoundPlayer().isSoundMute());
 		musicMuteItem.setSelected(desktop.getSoundPlayer().isMusicMute());
 		effectMuteItem.setSelected(desktop.getSoundPlayer().isSoundMute());
